Improved task structure, added react icon, vars
